Update rename-all-pages-in-module.js
Enhanced debug mode to be able to test different tasks. 
Updated task completion message. 
Refactored variables that had default values of null.
Account for case when checking for user choice input.
Added feature to remove prefix or suffix.
Optimized the script so it doesn't ask Canvas to save when no updates are made.